+++ gui/dsiprouter.py
@@ -198,7 +198,7 @@ def login():
         # if username valid, hash password and compare with stored password
         if form['username'] == settings.DSIP_USERNAME:
             if isinstance(settings.DSIP_PASSWORD, bytes):
-                pwcheck = Credentials.hashCreds(form['password'], settings.DSIP_PASSWORD[-Credentials.SALT_ENCODED_LEN:])
+                pwcheck = Credentials.hashCreds(form['password'], settings.DSIP_PASSWORD[-(Credentials.SALT_LEN * 2):])
             else:
                 pwcheck = form['password']
 
@@ -2366,54 +2366,101 @@ def syncSettings(new_fields={}, update_net=False):
         else:
             net_dict = {}
 
-        # sync settings from settings.py
-        if settings.LOAD_SETTINGS_FROM == 'file' or settings.DSIP_ID is None:
-            # need to grab any changes on disk b4 merging
-            reload(settings)
+        # need to grab any changes on disk b4 merging
+        reload(settings)
 
-            if settings.DSIP_ID is None:
-                with open('/etc/machine-id', 'r') as f:
-                    settings.DSIP_ID = Credentials.hashCreds(f.read().rstrip())
+        # if DSIP_ID is not set, generate it
+        if settings.DSIP_ID is None:
+            with open('/etc/machine-id', 'r') as f:
+                settings.DSIP_ID = Credentials.hashCreds(f.read().rstrip())
+        # if HOMER_ID is not set, generate it
+        if settings.HOMER_ID is None:
+            with open('/etc/machine-id', 'r') as f:
+                settings.HOMER_ID = int(Credentials.hashCreds(f.read().rstrip(), dklen=4)[0:8], 16)
+
+        # sync settings from settings.py
+        if settings.LOAD_SETTINGS_FROM == 'file':
 
             # format fields for DB
-            fields = OrderedDict(
-                [('DSIP_ID', settings.DSIP_ID), ('DSIP_CLUSTER_ID', settings.DSIP_CLUSTER_ID), ('DSIP_CLUSTER_SYNC', settings.DSIP_CLUSTER_SYNC),
-                 ('DSIP_PROTO', settings.DSIP_PROTO), ('DSIP_PORT', settings.DSIP_PORT), ('DSIP_USERNAME', settings.DSIP_USERNAME),
-                 ('DSIP_PASSWORD', settings.DSIP_PASSWORD), ('DSIP_IPC_PASS', settings.DSIP_IPC_PASS), ('DSIP_API_PROTO', settings.DSIP_API_PROTO),
-                 ('DSIP_API_PORT', settings.DSIP_API_PORT), ('DSIP_PRIV_KEY', settings.DSIP_PRIV_KEY), ('DSIP_PID_FILE', settings.DSIP_PID_FILE),
-                 ('DSIP_IPC_SOCK', settings.DSIP_IPC_SOCK), ('DSIP_UNIX_SOCK', settings.DSIP_UNIX_SOCK), ('DSIP_API_TOKEN', settings.DSIP_API_TOKEN),
-                 ('DSIP_LOG_LEVEL', settings.DSIP_LOG_LEVEL), ('DSIP_LOG_FACILITY', settings.DSIP_LOG_FACILITY),
-                 ('DSIP_SSL_KEY', settings.DSIP_SSL_KEY),
-                 ('DSIP_SSL_CERT', settings.DSIP_SSL_CERT), ('DSIP_SSL_CA', settings.DSIP_SSL_CA), ('DSIP_SSL_EMAIL', settings.DSIP_SSL_EMAIL),
-                 ('DSIP_CERTS_DIR', settings.DSIP_CERTS_DIR), ('VERSION', settings.VERSION), ('DEBUG', settings.DEBUG),
-                 ('ROLE', settings.ROLE), ('GUI_INACTIVE_TIMEOUT', settings.GUI_INACTIVE_TIMEOUT), ('KAM_DB_HOST', settings.KAM_DB_HOST),
-                 ('KAM_DB_DRIVER', settings.KAM_DB_DRIVER), ('KAM_DB_TYPE', settings.KAM_DB_TYPE), ('KAM_DB_PORT', settings.KAM_DB_PORT),
-                 ('KAM_DB_NAME', settings.KAM_DB_NAME), ('KAM_DB_USER', settings.KAM_DB_USER), ('KAM_DB_PASS', settings.KAM_DB_PASS),
-                 ('KAM_KAMCMD_PATH', settings.KAM_KAMCMD_PATH), ('KAM_CFG_PATH', settings.KAM_CFG_PATH),
-                 ('KAM_TLSCFG_PATH', settings.KAM_TLSCFG_PATH),
-                 ('RTP_CFG_PATH', settings.RTP_CFG_PATH),
-                 ('FLT_CARRIER', settings.FLT_CARRIER), ('FLT_PBX', settings.FLT_PBX), ('FLT_MSTEAMS', settings.FLT_MSTEAMS),
-                 ('FLT_OUTBOUND', settings.FLT_OUTBOUND), ('FLT_INBOUND', settings.FLT_INBOUND), ('FLT_LCR_MIN', settings.FLT_LCR_MIN),
-                 ('FLT_FWD_MIN', settings.FLT_FWD_MIN), ('DEFAULT_AUTH_DOMAIN', settings.DEFAULT_AUTH_DOMAIN),
-                 ('TELEBLOCK_GW_ENABLED', settings.TELEBLOCK_GW_ENABLED),
-                 ('TELEBLOCK_GW_IP', settings.TELEBLOCK_GW_IP), ('TELEBLOCK_GW_PORT', settings.TELEBLOCK_GW_PORT),
-                 ('TELEBLOCK_MEDIA_IP', settings.TELEBLOCK_MEDIA_IP),
-                 ('TELEBLOCK_MEDIA_PORT', settings.TELEBLOCK_MEDIA_PORT), ('FLOWROUTE_ACCESS_KEY', settings.FLOWROUTE_ACCESS_KEY),
-                 ('FLOWROUTE_SECRET_KEY', settings.FLOWROUTE_SECRET_KEY),
-                 ('FLOWROUTE_API_ROOT_URL', settings.FLOWROUTE_API_ROOT_URL), ('HOMER_HEP_HOST', settings.HOMER_HEP_HOST),
-                 ('HOMER_HEP_PORT', settings.HOMER_HEP_PORT),
-                 ('IPV6_ENABLED', settings.IPV6_ENABLED), ('INTERNAL_IP_ADDR', settings.INTERNAL_IP_ADDR),
-                 ('INTERNAL_IP_NET', settings.INTERNAL_IP_NET), ('INTERNAL_IP6_ADDR', settings.INTERNAL_IP6_ADDR),
-                 ('INTERNAL_IP6_NET', settings.INTERNAL_IP6_NET),
-                 ('EXTERNAL_IP_ADDR', settings.EXTERNAL_IP_ADDR), ('EXTERNAL_IP6_ADDR', settings.EXTERNAL_IP6_ADDR),
-                 ('EXTERNAL_FQDN', settings.EXTERNAL_FQDN),
-                 ('UPLOAD_FOLDER', settings.UPLOAD_FOLDER), ('MAIL_SERVER', settings.MAIL_SERVER),
-                 ('MAIL_PORT', settings.MAIL_PORT), ('MAIL_USE_TLS', settings.MAIL_USE_TLS), ('MAIL_USERNAME', settings.MAIL_USERNAME),
-                 ('MAIL_PASSWORD', settings.MAIL_PASSWORD), ('MAIL_ASCII_ATTACHMENTS', settings.MAIL_ASCII_ATTACHMENTS),
-                 ('MAIL_DEFAULT_SENDER', settings.MAIL_DEFAULT_SENDER), ('MAIL_DEFAULT_SUBJECT', settings.MAIL_DEFAULT_SUBJECT),
-                 ('DSIP_CORE_LICENSE', settings.DSIP_CORE_LICENSE), ('DSIP_STIRSHAKEN_LICENSE', settings.DSIP_STIRSHAKEN_LICENSE),
-                 ('DSIP_TRANSNEXUS_LICENSE', settings.DSIP_TRANSNEXUS_LICENSE), ('DSIP_MSTEAMS_LICENSE', settings.DSIP_MSTEAMS_LICENSE)]
-            )
+            fields = OrderedDict([
+                ('DSIP_ID', settings.DSIP_ID),
+                ('DSIP_CLUSTER_ID', settings.DSIP_CLUSTER_ID),
+                ('DSIP_CLUSTER_SYNC', settings.DSIP_CLUSTER_SYNC),
+                ('DSIP_PROTO', settings.DSIP_PROTO),
+                ('DSIP_PORT', settings.DSIP_PORT),
+                ('DSIP_USERNAME', settings.DSIP_USERNAME),
+                ('DSIP_PASSWORD', settings.DSIP_PASSWORD),
+                ('DSIP_IPC_PASS', settings.DSIP_IPC_PASS),
+                ('DSIP_API_PROTO', settings.DSIP_API_PROTO),
+                ('DSIP_API_PORT', settings.DSIP_API_PORT),
+                ('DSIP_PRIV_KEY', settings.DSIP_PRIV_KEY),
+                ('DSIP_PID_FILE', settings.DSIP_PID_FILE),
+                ('DSIP_IPC_SOCK', settings.DSIP_IPC_SOCK),
+                ('DSIP_UNIX_SOCK', settings.DSIP_UNIX_SOCK),
+                ('DSIP_API_TOKEN', settings.DSIP_API_TOKEN),
+                ('DSIP_LOG_LEVEL', settings.DSIP_LOG_LEVEL),
+                ('DSIP_LOG_FACILITY', settings.DSIP_LOG_FACILITY),
+                ('DSIP_SSL_KEY', settings.DSIP_SSL_KEY),
+                ('DSIP_SSL_CERT', settings.DSIP_SSL_CERT),
+                ('DSIP_SSL_CA', settings.DSIP_SSL_CA),
+                ('DSIP_SSL_EMAIL', settings.DSIP_SSL_EMAIL),
+                ('DSIP_CERTS_DIR', settings.DSIP_CERTS_DIR),
+                ('VERSION', settings.VERSION),
+                ('DEBUG', settings.DEBUG),
+                ('ROLE', settings.ROLE),
+                ('GUI_INACTIVE_TIMEOUT', settings.GUI_INACTIVE_TIMEOUT),
+                ('KAM_DB_HOST', settings.KAM_DB_HOST),
+                ('KAM_DB_DRIVER', settings.KAM_DB_DRIVER),
+                ('KAM_DB_TYPE', settings.KAM_DB_TYPE),
+                ('KAM_DB_PORT', settings.KAM_DB_PORT),
+                ('KAM_DB_NAME', settings.KAM_DB_NAME),
+                ('KAM_DB_USER', settings.KAM_DB_USER),
+                ('KAM_DB_PASS', settings.KAM_DB_PASS),
+                ('KAM_KAMCMD_PATH', settings.KAM_KAMCMD_PATH),
+                ('KAM_CFG_PATH', settings.KAM_CFG_PATH),
+                ('KAM_TLSCFG_PATH', settings.KAM_TLSCFG_PATH),
+                ('RTP_CFG_PATH', settings.RTP_CFG_PATH),
+                ('FLT_CARRIER', settings.FLT_CARRIER),
+                ('FLT_PBX', settings.FLT_PBX),
+                ('FLT_MSTEAMS', settings.FLT_MSTEAMS),
+                ('FLT_OUTBOUND', settings.FLT_OUTBOUND),
+                ('FLT_INBOUND', settings.FLT_INBOUND),
+                ('FLT_LCR_MIN', settings.FLT_LCR_MIN),
+                ('FLT_FWD_MIN', settings.FLT_FWD_MIN),
+                ('DEFAULT_AUTH_DOMAIN', settings.DEFAULT_AUTH_DOMAIN),
+                ('TELEBLOCK_GW_ENABLED', settings.TELEBLOCK_GW_ENABLED),
+                ('TELEBLOCK_GW_IP', settings.TELEBLOCK_GW_IP),
+                ('TELEBLOCK_GW_PORT', settings.TELEBLOCK_GW_PORT),
+                ('TELEBLOCK_MEDIA_IP', settings.TELEBLOCK_MEDIA_IP),
+                ('TELEBLOCK_MEDIA_PORT', settings.TELEBLOCK_MEDIA_PORT),
+                ('FLOWROUTE_ACCESS_KEY', settings.FLOWROUTE_ACCESS_KEY),
+                ('FLOWROUTE_SECRET_KEY', settings.FLOWROUTE_SECRET_KEY),
+                ('FLOWROUTE_API_ROOT_URL', settings.FLOWROUTE_API_ROOT_URL),
+                ('HOMER_ID', settings.HOMER_ID),
+                ('HOMER_HEP_HOST', settings.HOMER_HEP_HOST),
+                ('HOMER_HEP_PORT', settings.HOMER_HEP_PORT),
+                ('IPV6_ENABLED', settings.IPV6_ENABLED),
+                ('INTERNAL_IP_ADDR', settings.INTERNAL_IP_ADDR),
+                ('INTERNAL_IP_NET', settings.INTERNAL_IP_NET),
+                ('INTERNAL_IP6_ADDR', settings.INTERNAL_IP6_ADDR),
+                ('INTERNAL_IP6_NET', settings.INTERNAL_IP6_NET),
+                ('EXTERNAL_IP_ADDR', settings.EXTERNAL_IP_ADDR),
+                ('EXTERNAL_IP6_ADDR', settings.EXTERNAL_IP6_ADDR),
+                ('EXTERNAL_FQDN', settings.EXTERNAL_FQDN),
+                ('UPLOAD_FOLDER', settings.UPLOAD_FOLDER),
+                ('MAIL_SERVER', settings.MAIL_SERVER),
+                ('MAIL_PORT', settings.MAIL_PORT),
+                ('MAIL_USE_TLS', settings.MAIL_USE_TLS),
+                ('MAIL_USERNAME', settings.MAIL_USERNAME),
+                ('MAIL_PASSWORD', settings.MAIL_PASSWORD),
+                ('MAIL_ASCII_ATTACHMENTS', settings.MAIL_ASCII_ATTACHMENTS),
+                ('MAIL_DEFAULT_SENDER', settings.MAIL_DEFAULT_SENDER),
+                ('MAIL_DEFAULT_SUBJECT', settings.MAIL_DEFAULT_SUBJECT),
+                ('DSIP_CORE_LICENSE', settings.DSIP_CORE_LICENSE),
+                ('DSIP_STIRSHAKEN_LICENSE', settings.DSIP_STIRSHAKEN_LICENSE),
+                ('DSIP_TRANSNEXUS_LICENSE', settings.DSIP_TRANSNEXUS_LICENSE),
+                ('DSIP_MSTEAMS_LICENSE', settings.DSIP_MSTEAMS_LICENSE),
+            ])
 
             fields.update(new_fields)
             fields.update(net_dict)
@@ -2614,13 +2661,14 @@ def teardown():
     # TODO: multiprocessing and bjoern modules try to handle this in the ExitException handlers
     #       marked for review...
     try:
-       os.remove(settings.DSIP_IPC_SOCK)
+        os.remove(settings.DSIP_IPC_SOCK)
     except:
-       pass
+        pass
     try:
-       os.remove(settings.DSIP_UNIX_SOCK)
+        os.remove(settings.DSIP_UNIX_SOCK)
     except:
-       pass
+        pass
+
 
 def main():
     try:
@@ -2639,6 +2687,7 @@ def main():
             os._exit(0)
         # an exception bubbled up, allow caller to handler it
         raise
+
 
 if __name__ == "__main__":
     main()